// scripts/types.d.ts
// Auto-generated by scripts/generate-types.ts
export interface Profile {
    email:    string;
    fullName: string;
    location: string;
    name:     string;
    skills:   Skills;
    socials:  Socials;
    title:    string;
    website:  string;
}

export interface Skills {
    frameworks:     Framework[];
    infrastructure: Infrastructure[];
    languages:      Language[];
}

export interface Framework {
    confidence: number;
    name:       string;
}

export interface Infrastructure {
    confidence: number;
    name:       string;
}

export interface Language {
    confidence: number;
    name:       string;
}

export interface Socials {
    github:         string;
    githubUsername: string;
    linkedin?:      string;
}


export interface WakaTime {
    data: Data;
}

export interface Data {
    categories:                                            Category[];
    daily_average:                                         number;
    daily_average_including_other_language:                number;
    days_including_holidays:                               number;
    dependencies:                                          any[];
    editors:                                               Category[];
    end:                                                   Date;
    holidays:                                              number;
    human_readable_daily_average:                          string;
    human_readable_daily_average_including_other_language: string;
    human_readable_range:                                  string;
    human_readable_total:                                  string;
    human_readable_total_including_other_language:         string;
    id:                                                    string;
    is_already_updating:                                   boolean;
    is_coding_activity_visible:                            boolean;
    is_including_today:                                    boolean;
    is_other_usage_visible:                                boolean;
    is_stuck:                                              boolean;
    is_up_to_date:                                         boolean;
    languages:                                             Category[];
    operating_systems:                                     Category[];
    percent_calculated:                                    number;
    range:                                                 string;
    start:                                                 Date;
    status:                                                string;
    timeout:                                               number;
    total_seconds:                                         number;
    total_seconds_including_other_language:                number;
    user_id:                                               string;
    username:                                              string;
    writes_only:                                           boolean;
}

export interface Category {
    digital:       string;
    hours:         number;
    minutes:       number;
    name:          string;
    percent:       number;
    text:          string;
    total_seconds: number;
}